moved validation logic to its own file
